requirements.txt
fastmcp
uvicorn
requests
python-dotenv
psycopg2
psycopg2-binary